feat: remove torch deps
